cleaning of control variable definition
