+++ CMGTools/TTHAnalysis/python/analyzers/ttHPhotonAnalyzerSusy.py
@@ -94,6 +94,59 @@ class ttHPhotonAnalyzerSusy( Analyzer ):
             gen = match[gamma]
             gamma.mcMatchId = 1 if gen else 0
 
+### MM
+    def makeGammaObjects(self, event):
+
+        import ROOT
+
+        ## with Central Jets
+        gamma_objects25 = [ j for j in event.gamma_cleanJets if j.pt() > 25 ] + event.selectedLeptons
+        gamma_objects30 = [ j for j in event.gamma_cleanJets if j.pt() > 30 ] + event.selectedLeptons
+        gamma_objects40  = [ j for j in event.gamma_cleanJetsAll if j.pt() > 40 and abs(j.eta()) < 2.5 ] + [ l for l in event.selectedLeptons if l.pt() > 10 and abs(l.eta()) < 2.5 ]
+        gamma_objects40j = [ j for j in event.gamma_cleanJets if j.pt() > 40 ]
+
+        event.gamma_htJet25 = sum([x.pt() for x in gamma_objects25])
+        event.gamma_mhtJet25vec = ROOT.reco.Particle.LorentzVector(-1.*(sum([x.px() for x in gamma_objects25])) , -1.*(sum([x.py() for x in gamma_objects25])), 0, 0 )
+        event.gamma_mhtPhiJet25 = event.mhtJet25vec.phi()
+        event.gamma_mhtJet25 = event.mhtJet25vec.pt()
+
+        event.gamma_htJet30 = sum([x.pt() for x in gamma_objects30])
+        event.gamma_mhtJet30vec = ROOT.reco.Particle.LorentzVector(-1.*(sum([x.px() for x in gamma_objects30])) , -1.*(sum([x.py() for x in gamma_objects30])), 0, 0 )
+        event.gamma_mhtJet30 = event.mhtJet30vec.pt()
+        event.gamma_mhtPhiJet30 = event.mhtJet30vec.phi()
+
+        event.gamma_htJet40 = sum([x.pt() for x in gamma_objects40])
+        event.gamma_mhtJet40vec = ROOT.reco.Particle.LorentzVector(-1.*(sum([x.px() for x in gamma_objects40])) , -1.*(sum([x.py() for x in gamma_objects40])), 0, 0 )
+        event.gamma_mhtJet40 = event.gamma_mhtJet40vec.pt()
+        event.gamma_mhtPhiJet40 = event.gamma_mhtJet40vec.phi()
+
+        event.gamma_htJet40j = sum([x.pt() for x in gamma_objects40j])
+        event.gamma_mhtJet40jvec = ROOT.reco.Particle.LorentzVector(-1.*(sum([x.px() for x in gamma_objects40j])) , -1.*(sum([x.py() for x in gamma_objects40j])), 0, 0 )
+        event.gamma_mhtJet40j = event.gamma_mhtJet40jvec.pt()
+        event.gamma_mhtPhiJet40j = event.gamma_mhtJet40jvec.phi()
+
+         # MET + photon
+        event.gamma_met = ROOT.reco.Particle.LorentzVector( event.met.px(), event.met.py(), 0, 0 )
+        event.gamma_metNoPU = ROOT.reco.Particle.LorentzVector( event.metNoPU.px(), event.metNoPU.py(), 0, 0 )
+        for gamma in event.loosePhotonsCentral:
+            event.gamma_met = ROOT.reco.Particle.LorentzVector( event.gamma_met.px() + gamma.px(), event.gamma_met.py() + gamma.py() , 0, 0 )
+            event.gamma_metNoPU = ROOT.reco.Particle.LorentzVector( event.gamma_metNoPU.px() + gamma.px(), event.gamma_metNoPU.py() + gamma.py() , 0, 0 )
+
+        # look for minimal deltaPhi between MET and four leading jets with pt>40 and eta<2.4
+        event.gamma_deltaPhiMin = 999.
+        for n,j in enumerate(gamma_objects40j):
+            if n>3:  break
+            thisDeltaPhi = abs( deltaPhi( j.phi(), event.gamma_met.phi() ) )
+            if thisDeltaPhi < event.gamma_deltaPhiMin : event.gamma_deltaPhiMin = thisDeltaPhi
+
+        # absolute value of the vectorial difference between met and mht
+        gamma_diffMetMht_had_vec = ROOT.reco.Particle.LorentzVector(event.gamma_mhtJet40jvec.px()-event.gamma_met.px(), event.gamma_mhtJet40jvec.py()-event.gamma_met.py(), 0, 0 )
+        event.gamma_diffMetMht_had = sqrt( gamma_diffMetMht_had_vec.px()*gamma_diffMetMht_had_vec.px() + gamma_diffMetMht_had_vec.py()*gamma_diffMetMht_had_vec.py() )
+
+        gamma_diffMetMht_vec = ROOT.reco.Particle.LorentzVector(event.gamma_mhtJet40vec.px()-event.gamma_met.px(), event.gamma_mhtJet40vec.py()-event.gamma_met.py(), 0, 0 )
+        event.gamma_diffMetMht = sqrt( gamma_diffMetMht_vec.px()*gamma_diffMetMht_vec.px() + gamma_diffMetMht_vec.py()*gamma_diffMetMht_vec.py() )
+
+
     def printInfo(self, event):
         print '----------------'
         if len(event.selectedPhotons)>0:
@@ -114,6 +167,7 @@ class ttHPhotonAnalyzerSusy( Analyzer ):
         self.readCollections( iEvent )
         #call the photons functions
         self.makePhotons(event)
+        self.makeGammaObjects(event)
         
 #        self.printInfo(event)
